chore(node-repository): merge `InfraApplicationRedeployer` into `InfraDeployer`
Allow `readied()` method to be used in external node-repository maintainers.
