+++ config-provisioning/src/main/java/com/yahoo/config/provision/InfraDeployer.java
@@ -7,8 +7,9 @@ import java.util.Optional;
  * A deployer is used to deploy infrastructure applications.
  *
  * @author freva
+ * @author bjorncs
  */
-public interface InfraDeployer {
+public interface InfraDeployer extends AutoCloseable {
 
     /**
      * @param application the infrastructure application to be deployed
@@ -18,5 +19,10 @@ public interface InfraDeployer {
 
     /** Deploys all supported infrastructure applications in this zone. */
     void activateAllSupportedInfraApplications(boolean propagateException);
+
+    /** Marks the specified node type as ready and triggers redeployment of its corresponding infrastructure application */
+    void readied(NodeType type);
+
+    void close();
 
 }
