+++ node-repository/src/main/java/com/yahoo/vespa/hosted/provision/provisioning/InfraDeployerImpl.java
@@ -2,8 +2,11 @@
 package com.yahoo.vespa.hosted.provision.provisioning;
 
 import ai.vespa.http.DomainName;
+import com.yahoo.component.AbstractComponent;
 import com.yahoo.component.Version;
 import com.yahoo.component.annotation.Inject;
+import com.yahoo.concurrent.DaemonThreadFactory;
+import com.yahoo.concurrent.UncheckedTimeoutException;
 import com.yahoo.config.provision.ActivationContext;
 import com.yahoo.config.provision.ApplicationId;
 import com.yahoo.config.provision.ApplicationTransaction;
@@ -13,7 +16,11 @@ import com.yahoo.config.provision.HostSpec;
 import com.yahoo.config.provision.InfraDeployer;
 import com.yahoo.config.provision.NodeType;
 import com.yahoo.config.provision.Provisioner;
+import com.yahoo.transaction.Mutex;
 import com.yahoo.transaction.NestedTransaction;
+import com.yahoo.vespa.applicationmodel.InfrastructureApplication;
+import com.yahoo.vespa.hosted.provision.Node;
+import com.yahoo.vespa.hosted.provision.NodeList;
 import com.yahoo.vespa.hosted.provision.NodeRepository;
 import com.yahoo.vespa.hosted.provision.maintenance.InfrastructureVersions;
 import com.yahoo.vespa.service.monitor.DuperModelInfraApi;
@@ -22,14 +29,29 @@ import com.yahoo.vespa.service.monitor.InfraApplicationApi;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
+import java.util.Set;
+import java.util.concurrent.ConcurrentSkipListSet;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.TimeUnit;
+import java.util.function.Function;
+import java.util.function.Supplier;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 
+import static java.util.logging.Level.FINE;
+import static java.util.logging.Level.INFO;
+import static java.util.logging.Level.WARNING;
+
 /**
+ * Performs on-demand redeployment of the {@link InfrastructureApplication}s, to minimise time between
+ * host provisioning for a deployment completing, and deployment of its application containers succeeding.
+ *
  * @author freva
+ * @author bjorncs
  */
-public class InfraDeployerImpl implements InfraDeployer {
+public class InfraDeployerImpl extends AbstractComponent implements InfraDeployer, AutoCloseable {
 
     private static final Logger logger = Logger.getLogger(InfraDeployerImpl.class.getName());
 
@@ -37,13 +59,24 @@ public class InfraDeployerImpl implements InfraDeployer {
     private final Provisioner provisioner;
     private final InfrastructureVersions infrastructureVersions;
     private final DuperModelInfraApi duperModel;
+    private final ExecutorService executor = Executors.newSingleThreadExecutor(new DaemonThreadFactory("infra-application-redeployer-"));
+    private final Set<InfrastructureApplication> readiedTypes = new ConcurrentSkipListSet<>();
+    private final Function<ApplicationId, Mutex> locks;
+    private final Supplier<NodeList> nodes;
 
     @Inject
     public InfraDeployerImpl(NodeRepository nodeRepository, Provisioner provisioner, DuperModelInfraApi duperModel) {
+        this(nodeRepository, provisioner, duperModel, nodeRepository.applications()::lockMaintenance, nodeRepository.nodes()::list);
+    }
+
+    InfraDeployerImpl(NodeRepository nodeRepository, Provisioner provisioner, DuperModelInfraApi duperModel,
+                      Function<ApplicationId, Mutex> locks, Supplier<NodeList> nodes) {
         this.nodeRepository = nodeRepository;
         this.provisioner = provisioner;
         this.infrastructureVersions = nodeRepository.infrastructureVersions();
         this.duperModel = duperModel;
+        this.locks = locks;
+        this.nodes = nodes;
     }
 
     @Override
@@ -72,6 +105,70 @@ public class InfraDeployerImpl implements InfraDeployer {
 
         duperModel.infraApplicationsIsNowComplete();
     }
+
+    @Override
+    public void readied(NodeType type) {
+        applicationOf(type).ifPresent(this::readied);
+    }
+
+    @Override
+    public void close() {
+        executor.shutdown();
+        try {
+            if (executor.awaitTermination(10, TimeUnit.SECONDS)) return;
+            logger.log(WARNING, "Redeployer did not shut down within 10 seconds");
+        }
+        catch (InterruptedException e) {
+            Thread.currentThread().interrupt();
+        }
+        executor.shutdownNow();
+    }
+
+    @Override public void deconstruct() { close(); }
+
+    private void readied(InfrastructureApplication application) {
+        if (application == null) return;
+        if (readiedTypes.add(application)) executor.execute(() -> checkAndRedeploy(application));
+    }
+
+    private void checkAndRedeploy(InfrastructureApplication application) {
+        if ( ! readiedTypes.remove(application)) return;
+        try (Mutex lock = locks.apply(application.id())) {
+            if (application.nodeType().isHost() && nodes.get().state(Node.State.ready).nodeType(application.nodeType()).isEmpty()) return;
+            logger.log(FINE, () -> "Redeploying " + application.id() + " after completing provisioning for " + application.name());
+            try {
+                getDeployment(application.id()).ifPresent(Deployment::activate);
+                childOf(application).ifPresent(this::readied);
+            }
+            catch (RuntimeException e) {
+                logger.log(INFO, "Failed redeploying " + application.id() + ", will be retried by maintainer", e);
+            }
+        }
+        catch (UncheckedTimeoutException collision) {
+            readied(application);
+        }
+    }
+
+    private static Optional<InfrastructureApplication> applicationOf(NodeType type) {
+        return switch (type) {
+            case host -> Optional.of(InfrastructureApplication.TENANT_HOST);
+            case confighost -> Optional.of(InfrastructureApplication.CONFIG_SERVER_HOST);
+            case config -> Optional.of(InfrastructureApplication.CONFIG_SERVER);
+            case controllerhost -> Optional.of(InfrastructureApplication.CONTROLLER_HOST);
+            case controller -> Optional.of(InfrastructureApplication.CONTROLLER);
+            case proxyhost -> Optional.of(InfrastructureApplication.PROXY_HOST);
+            default -> Optional.empty();
+        };
+    }
+
+    private static Optional<InfrastructureApplication> childOf(InfrastructureApplication application) {
+        return switch (application) {
+            case CONFIG_SERVER_HOST -> Optional.of(InfrastructureApplication.CONFIG_SERVER);
+            case CONTROLLER_HOST -> Optional.of(InfrastructureApplication.CONTROLLER);
+            default -> Optional.empty();
+        };
+    }
+
 
     private class InfraDeployment implements Deployment {
